Photo upload refactor. Put photo uploading into seperate component.
